fix rendering events on the page
