+++ src/partials/header.html
@@ -2,7 +2,7 @@
     <div class="container header-position">
         <h1 class="header-sleng">FIND BEST EVENTS <br> AROUND THE WORLD</h1>
           <form class="area-form">
-            <input class="earch-event-area" type="text" id="event" name="query_event" autocomplete="off" placeholder="Start searching" required pattern="^[a-zA-Z]+$">
+            <input class="search-event-area" type="text" id="event" name="query_event" autocomplete="off" placeholder="Start searching" required pattern="^[a-zA-Z]+$">
             <input class="search-country-area" type="text" id="country" name="query_country" autocomplete="off" placeholder="Choose country" required pattern="^[a-zA-Z]+$">
           </form>
     </div>
