+++ src/templation/gallery.hbs
@@ -1,4 +1,4 @@
-{{#each this}}
+  {{#each this}}
   <li class='event-card' id={{id}}>
 
     <img class='event-image' src='{{images.3.url}}' alt='{{name}}' width=350 />
@@ -9,4 +9,4 @@
         <a href="{{_embedded.venues.0.location.latitude}}" class="link" target="_blank"> <span class="material-icons cards-icons">room</span> {{_embedded.venues.0.name}}</a>
     </p>
   </li>
-{{/each}}
+  {{/each}}
